% GenometriCorrelation project evaluating two interval markups genomewide independence. 
% (c) 2010-2014 Alexander Favorov, Loris Mularoni, Yulia Medvedeva, 
%               Harris A. Jaffee, Ekaterina V. Zhuravleva, Leslie M. Cope, 
%               Andrey A. Mironov, Vsevolod J. Makeev, Sarah J. Wheelan.
%
\name{GenometriCorrResult-class}
\Rdversion{1.1}
\docType{class}
\alias{GenometriCorrResult-class}
\alias{show,GenometriCorrResult-method}
\alias{graphical.report,GenometriCorrResult-method}
\alias{graphical.report}
\alias{visualize,GenometriCorrResult-method}
\alias{visualize}

\title{The class contains the results of the GenometriCorrelation function run}

\description{The \code{GenometriCorrResult-class} is generated by \code{\link{GenometriCorrelation}} or \code{GenometriCorr:::run.config}; it is inherited from a \code{namedList} that contains the the list of results. Three mechanisms, \code{print}, \code{graphical.report}, and \code{visualize} create text and/or graphics that summarize the results in a human-readable way.}

\section{Creating Objects}{
\code{new('GenometriCorrConfig',src)}\cr
The object is generated by \code{\link{GenometriCorrelation}} or \code{GenometriCorr:::run.config};
}

\section{Slots}{
  \describe{
    \item{\code{.Data}:}{Object of class \code{"list"} ~~ }
    \item{\code{config}:}{Object of class \code{"GenometriCorrConfig"} ~~ }
    \item{\code{names}:}{Object of class \code{"character"} ~~ }
  }
}

\section{Slots}{
	\describe{
		\item{\code{.Data}}{The result of the \code{\link{GenometriCorrelation}} call. It is a list and the \code{names} slot contains the names of the members of the result list.} 
		\item{\code{config}}{The config in an instanse of \code{\link{GenometriCorrConfig-class}} class that stores the paprameters that were used in the \code{\link{GenometriCorrelation}} run that gave the results. }
    \item{\code{names}:}{Object of class \code{"character"}; names of the resulting list; it saves of the spaces (chromosomes) used for the calculations. It is inherited from "\linkS4class{namedList}"}
	}
}
\section{Extends}{
Class \code{"\linkS4class{namedList}"}, directly.
Class \code{"\linkS4class{list}"}, by class "namedList", distance 2.
Class \code{"\linkS4class{vector}"}, by class "namedList", distance 3.
}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "GenometriCorrResult")}: writes the results as plain text.}
    \item{graphical.report}{\code{signature(x = "GenometriCorrResult")}:\cr
			\code{graphical.report(result,pdffile,show.all,show.chromosomes,trustname,make.new)}{plots the results as a pdf, named \code{pdffile}. If a filename is not given, it creates the filename from the original datafile names. \code{show.all} is a boolean, determining whether to show graphcs for all chromosomes as well as the summary statistics for all data (default is \code{TRUE}). If \code{FALSE}, only the summary statistics for all data will appear in the output. \code{chromosomes.to.show} lists which individual chromosome results should be in the output; in this case the summary statistics for all data will not be shown. Default is (\code{""}).}If \code{trustname} is \code{TRUE} it does not try to add a \code{.pdf} extension to the filename supplied. 
		}
    \item{visualize}{\code{signature(x = "GenometriCorrResult")}:\cr
			\code{visualize(result,pdffile,show.all,show.chromosomes,trustname,make.new)}behaves much like the \code{graphical.report} function but generates a pdf with a full graphical overview. \code{pdffile, show.all}, and \code{chromosomes.to.show} behave as for \code{graphical.report}. \code{style} is a string with the name of the color style. Currently, only two are available: \code{"bleu-white-red"} (default) and \code{"rainbow"} (the old one). The graphical overview depicts both data density and the significance of trends in the spatial relationship between query and reference data.
		}
	 }
}

\references{
%% ~put references to the literature/web site here ~
\url{http://genometricorr.sourceforge.net/}
}

\author{Alexander Favorov \email{favorov@sensi.org}, Loris Mularoni, Leslie M. Cope, Andrey A. Mironov,  Yulia Medvedeva, Vsevolod J. Makeev, Sarah J. Wheelan}

\seealso{
	The \code{\link{GenometriCorrResult-class}} class; the \code{\link{GenometriCorr}} documentation and the \href{../doc/GenometriCorr.pdf}{GenometriCorr package vingnette}.
}

\examples{

library("GenometriCorr")

cpgis_bed<-system.file("extdata", "UCSCcpgis_hg19.bed", package = "GenometriCorr");

refseq_bed<-system.file("extdata", "UCSCrefseqgenes_hg19.bed", package = "GenometriCorr");

config<-new("GenometriCorrConfig",system.file("extdata","template-config.ini", package = "GenometriCorr"))
#creating the new config from the template-config.ini

config$tests$ecdf.area.permut.number=10
config$tests$mean.distance.permut.number=10
config$tests$jaccard.measure.permut.number=10

#sets all permutation numbers to 10

config$options$showTkProgressBar=FALSE
#supresss the Tk progress bar

print(config)
#printing the changed file back, just to see it 

result<-run.config(config,query=cpgis_bed,reference=refseq_bed)
#runnig GenometriCorrelation with the parameters from config
#and query and reference file names as given

print(result)

graphical.report(result)
}
\keyword{classes}
