% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseTwoIRanges.R
\name{VisualiseTwoGRanges}
\alias{VisualiseTwoGRanges}
\title{VisualiseTwoGRanges}
\usage{
VisualiseTwoGRanges(
  grA,
  grB,
  nameA = "RangesA",
  nameB = "RangesB",
  title = NA,
  pdf = NULL,
  close.device = NULL
)
}
\arguments{
\item{grA}{First \code{GRanges} object to be vsualised.}

\item{grB}{Second \code{GRanges} object to be vsualised.}

\item{nameA}{Name of \code{grA}, default is 'RangesA'}

\item{nameB}{Name of \code{grB}, default is 'RangesB'}

\item{title}{Title, printed at the top of the plot.}

\item{pdf}{Name of a file to which the image should be written. If \code{pdf=""} the filename is constructed from \code{nameA} and \code{nameB}. The suffix ".pdf" will be appended if not included. If \code{NULL}, no pdf is opened and an x11 window is raised up as by \code{plot}.}

\item{close.device}{Whether to close the plot device after writing the image. A \code{FALSE} setting allows multiple images to be written to the same pdf file, but then eventually closing the device is up to the user. Default is \code{NULL} that means: close an x11, do not close a pdf.}
}
\description{
VisualiseTwoGRanges is a function that displays the markups that are contatined two GRanges, in a multi-page pdf, page per chromosome, by \code{\link{VisualiseTwoIRanges}} for aech chromosome. The intent is to show large-scale relationships between the two markups. All the chromosome lenght data is provided by the GRanges.
}
\references{
\href{http://genometricorr.sourceforge.net/}{GenometriCorr home}
}
\seealso{
The \code{\link{GenometriCorr}} documentation and vignette.
}
\author{
Alexander Favorov \email{favorov@sensi.org}, Loris Mularoni, Yulia Medvedeva, Harris A. Jaffee, Ekaterina V. Zhuravleva, Leslie M. Cope, Andrey A. Mironov, Vsevolod J. Makeev, Sarah J. Wheelan.
}
\concept{GenometriCorr 2-range visualisations}
\keyword{hplot}
