% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseTwoIRanges.R
\name{VisualiseTwoIRanges}
\alias{VisualiseTwoIRanges}
\title{VisualiseTwoIRanges}
\usage{
VisualiseTwoIRanges(irA, irB, start = 1, end = NA, nameA = "RangesA",
  nameB = "RangesB", chrom_length = NA, title = NA, pdf = NULL,
  close.device = NULL)
}
\arguments{
\item{irA}{First \code{IRanges} object to be vsualised.}

\item{irB}{Second \code{IRanges} object to be vsualised.}

\item{start}{Start of the visualisation band.}

\item{end}{End of the visualisation band.}

\item{nameA}{Name of \code{irA}, default is 'RangesA'}

\item{nameB}{Name of \code{irB}, default is 'RangesB'}

\item{chrom_length}{The length of the chromosome spanned by \code{irA} and \code{irB}}

\item{title}{Title, printed at the top of the plot.}

\item{pdf}{Name of a file to which the image should be written. If \code{pdf=""} the filename is constructed from \code{nameA} and \code{nameB}. The suffix ".pdf" will be appended if not included. If \code{NULL}, no pdf is opened and an x11 window is raised up as by \code{plot}.}

\item{close.device}{Whether to close the plot device after writing the image. A \code{FALSE} setting allows multiple images to be written to the same pdf file, but then eventually closing the device is up to the user. Default is \code{NULL} that means: close an x11, do not close a pdf.}
}
\description{
VisualiseTwoIRanges is a function that displays the intervals of two IRanges, one above the other, in different colors, along a chromosome or subset of a chromosome. The intent is to show large-scale relationships between the two IRanges.
}
\examples{

library('rtracklayer')
library('GenometriCorr')

cpgis<-as(import(system.file("extdata", "UCSCcpgis_hg19.bed", package = "GenometriCorr")),'RangedData');

refseq<-as(import(system.file("extdata", "UCSCrefseqgenes_hg19.bed", package = "GenometriCorr")),'RangedData');


human.chrom.length<-c(249250621,243199373,198022430,191154276,180915260,171115067,159138663,146364022,141213431,135534747,135006516,133851895,115169878,107349540,102531392,90354753,81195210,78077248,59128983,63025520,48129895,51304566,59373566,155270560)


names(human.chrom.length)<-c('chr1','chr2','chr3','chr4','chr5','chr6','chr7','chr8','chr9','chr10','chr11','chr12','chr13','chr14','chr15','chr16','chr17','chr18','chr19','chr20','chr21','chr22','chrY','chrX')

VisualiseTwoIRanges(cpgis['chr1']$ranges, refseq['chr1']$ranges,
	nameA='CpG Islands', nameB='RefSeq Genes',
	chrom_length=human.chrom.length[['chr1']],
	title="CpGIslands and RefGenes on chr1, hg19",
	pdf='CpGi_vs_RefSeq_genes_chr1_hg19')

}
\references{
\href{http://genometricorr.sourceforge.net/}{GenometriCorr home}
}
\seealso{
The \code{\link{GenometriCorr}} documentation and vignette.
}
\author{
Alexander Favorov \email{favorov@sensi.org}, Loris Mularoni, Yulia Medvedeva, Harris A. Jaffee, Ekaterina V. Zhuravleva, Leslie M. Cope, Andrey A. Mironov, Vsevolod J. Makeev, Sarah J. Wheelan.
}
\concept{GenometriCorr 2-range visualisation}
\keyword{hplot}
